requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
openai
google-generativeai
python-dotenv
